Epochs API sanity test
